print PIN-JSON in canonical form
Previously the propertys of the JSON-object were serialised in what seems to me randomly ordner. That caused every request to the PIN-API to return a JSON which, even for equal content, hat a different checksum causing the Tür-PI to rewrite it to the SD-Card.

With this patch the library always first writes the nickname and then writes the pin. Two subsequent requests to the PIN-API will yielt byte for byte the same file (if no pin was changed/added/removed) causing the write-cycle-reduction on the Tür-PI do work as expected.